# contracts/http/platform-control.openapi.yaml
openapi: 3.1.0
info:
  title: ERP Platform Control API
  version: 0.2.0
  description: Tenant capabilities, entitlements, quotas, metering, lifecycle jobs and SaaS governance.
paths:
...
  /api/platform-control/tenants/{tenantSlug}/entitlements:
    get:
      summary: List tenant entitlements with cursor pagination
      parameters:
        - in: path
          name: tenantSlug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Paginated entitlement list
  /api/platform-control/tenants/{tenantSlug}/entitlements/{capabilityKey}:
    put:
      summary: Upsert one entitlement
      parameters:
        - in: path
          name: tenantSlug
          required: true
          schema:
            type: string
        - in: path
          name: capabilityKey
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Upserted entitlement
...