# service-api/service-python/platform-control/app/server.py
"""Bootstrap HTTP do servico platform-control."""

from fastapi import FastAPI, Header, HTTPException
from fastapi.responses import JSONResponse

from app.config.settings import settings
from app.infrastructure.postgres import postgres_ready
from app.runtime import (
    build_go_live_readiness,
    build_lifecycle_preview,
    build_lifecycle_readiness,
    build_usage_summary,
    bulk_upsert_entitlements,
    bulk_upsert_quotas,
    create_go_live_rollout,
    create_lifecycle_job,
    create_metering_snapshot,
    get_go_live_rollout,
    get_lifecycle_job,
    list_blocks,
    list_capability_catalog,
    list_entitlements_page,
    list_go_live_rollouts,
    list_lifecycle_jobs_page,
    list_metering_page,
    list_provider_catalog,
    list_provider_defaults,
    list_quotas,
    transition_go_live_rollout,
    transition_lifecycle_job,
    upsert_block,
    upsert_entitlement,
    upsert_provider_default,
    upsert_quota,
)


app = FastAPI(title=settings.service_name)


@app.get("/health/live")
def live() -> dict:
    return {"service": settings.service_name, "status": "live"}


@app.get("/health/ready")
def ready() -> dict:
    return {"service": settings.service_name, "status": "ready"}


@app.get("/health/details")
def details() -> dict:
    dependencies = [
        {"name": "capability-registry", "status": "ready"},
        {"name": "entitlements", "status": "ready"},
        {"name": "metering", "status": "ready"},
        {"name": "provider-defaults", "status": "ready"},
        {"name": "tenant-lifecycle", "status": "ready"},
        {"name": "lifecycle-readiness", "status": "ready"},
        {"name": "go-live-control", "status": "ready"},
        {"name": "quotas", "status": "ready"},
        {"name": "tenant-blocks", "status": "ready"},
        {"name": "idempotency", "status": "ready"},
        {"name": "async-jobs", "status": "ready"},
    ]
    if settings.repository_driver == "postgres":
        dependencies.insert(0, {"name": "postgresql", "status": "ready" if postgres_ready() else "not_ready"})
    return {"service": settings.service_name, "status": "ready", "dependencies": dependencies}


@app.get("/api/platform-control/capabilities/catalog")
def get_catalog() -> list[dict]:
    return list_capability_catalog()


@app.get("/api/platform-control/providers/catalog")
def get_provider_catalog() -> list[dict]:
    return list_provider_catalog()


@app.get("/api/platform-control/tenants/{tenant_slug}/entitlements")
def entitlements(tenant_slug: str, cursor: str | None = None, limit: int = 50) -> dict:
    return list_entitlements_page(tenant_slug, cursor, limit)


@app.put("/api/platform-control/tenants/{tenant_slug}/entitlements/{capability_key}")
def put_entitlement(tenant_slug: str, capability_key: str, payload: dict) -> dict:
    try:
        return upsert_entitlement(tenant_slug, capability_key, payload)
    except ValueError as error:
        raise HTTPException(status_code=400, detail={"code": str(error), "message": "Entitlement payload is invalid."}) from error


@app.post("/api/platform-control/tenants/{tenant_slug}/entitlements/bulk")
def post_bulk_entitlements(tenant_slug: str, payload: dict) -> dict:
    return bulk_upsert_entitlements(tenant_slug, payload)


@app.get("/api/platform-control/tenants/{tenant_slug}/provider-defaults")
def provider_defaults(tenant_slug: str) -> dict:
    return {"tenantSlug": tenant_slug, "items": list_provider_defaults(tenant_slug)}


@app.put("/api/platform-control/tenants/{tenant_slug}/provider-defaults/{capability_key}")
def put_provider_default(tenant_slug: str, capability_key: str, payload: dict) -> dict:
    try:
        return upsert_provider_default(tenant_slug, capability_key, payload)
    except ValueError as error:
        raise HTTPException(status_code=400, detail={"code": str(error), "message": "Provider default payload is invalid."}) from error


@app.get("/api/platform-control/tenants/{tenant_slug}/quotas")
def quotas(tenant_slug: str) -> dict:
    return {"tenantSlug": tenant_slug, "items": list_quotas(tenant_slug)}


@app.put("/api/platform-control/tenants/{tenant_slug}/quotas/{metric_key}")
def put_quota(tenant_slug: str, metric_key: str, payload: dict) -> dict:
    try:
        return upsert_quota(tenant_slug, metric_key, payload)
    except ValueError as error:
        raise HTTPException(status_code=400, detail={"code": str(error), "message": "Quota payload is invalid."}) from error


@app.post("/api/platform-control/tenants/{tenant_slug}/quotas/bulk")
def post_bulk_quotas(tenant_slug: str, payload: dict) -> dict:
    return bulk_upsert_quotas(tenant_slug, payload)


@app.get("/api/platform-control/tenants/{tenant_slug}/blocks")
def blocks(tenant_slug: str) -> dict:
    return {"tenantSlug": tenant_slug, "items": list_blocks(tenant_slug)}


@app.put("/api/platform-control/tenants/{tenant_slug}/blocks/{block_key}")
def put_block(tenant_slug: str, block_key: str, payload: dict) -> dict:
    try:
        return upsert_block(tenant_slug, block_key, payload)
    except ValueError as error:
        raise HTTPException(status_code=400, detail={"code": str(error), "message": "Block payload is invalid."}) from error


@app.get("/api/platform-control/tenants/{tenant_slug}/metering")
def metering(tenant_slug: str, cursor: str | None = None, limit: int = 50) -> dict:
    return list_metering_page(tenant_slug, cursor, limit)


@app.post("/api/platform-control/tenants/{tenant_slug}/metering/snapshots")
def post_metering(tenant_slug: str, payload: dict) -> dict:
    try:
        return create_metering_snapshot(tenant_slug, payload)
    except ValueError as error:
        raise HTTPException(status_code=400, detail={"code": str(error), "message": "Metering snapshot payload is invalid."}) from error


@app.get("/api/platform-control/tenants/{tenant_slug}/usage-summary")
def usage_summary(tenant_slug: str) -> dict:
    return build_usage_summary(tenant_slug)


@app.get("/api/platform-control/tenants/{tenant_slug}/lifecycle/readiness")
def lifecycle_readiness(tenant_slug: str) -> dict:
    return build_lifecycle_readiness(tenant_slug)


@app.get("/api/platform-control/tenants/{tenant_slug}/go-live/readiness")
def go_live_readiness(tenant_slug: str) -> dict:
    return build_go_live_readiness(tenant_slug)


@app.get("/api/platform-control/tenants/{tenant_slug}/go-live/rollouts")
def go_live_rollouts(tenant_slug: str, cursor: str | None = None, limit: int = 50) -> dict:
    return list_go_live_rollouts(tenant_slug, cursor, limit)


@app.get("/api/platform-control/tenants/{tenant_slug}/go-live/rollouts/{public_id}")
def go_live_rollout_detail(tenant_slug: str, public_id: str) -> dict:
    rollout = get_go_live_rollout(tenant_slug, public_id)
    if rollout is None:
        raise HTTPException(status_code=404, detail={"code": "go_live_rollout_not_found", "message": "Go-live rollout was not found."})
    return rollout


@app.post("/api/platform-control/tenants/{tenant_slug}/go-live/rollouts")
def post_go_live_rollout(tenant_slug: str, payload: dict) -> dict:
    try:
        return create_go_live_rollout(tenant_slug, payload)
    except ValueError as error:
        raise HTTPException(status_code=400, detail={"code": str(error), "message": "Go-live rollout payload is invalid."}) from error


def _transition_go_live(tenant_slug: str, public_id: str, action: str, payload: dict | None = None) -> dict:
    try:
        return transition_go_live_rollout(tenant_slug, public_id, action, payload or {})
    except ValueError as error:
        if str(error) == "go_live_rollout_not_found":
            raise HTTPException(status_code=404, detail={"code": str(error), "message": "Go-live rollout was not found."}) from error
        raise HTTPException(status_code=400, detail={"code": str(error), "message": "Go-live rollout transition payload is invalid."}) from error


@app.post("/api/platform-control/tenants/{tenant_slug}/go-live/rollouts/{public_id}/start")
def start_go_live_rollout(tenant_slug: str, public_id: str, payload: dict | None = None) -> dict:
    return _transition_go_live(tenant_slug, public_id, "start", payload)


@app.post("/api/platform-control/tenants/{tenant_slug}/go-live/rollouts/{public_id}/complete")
def complete_go_live_rollout(tenant_slug: str, public_id: str, payload: dict | None = None) -> dict:
    return _transition_go_live(tenant_slug, public_id, "complete", payload)


@app.post("/api/platform-control/tenants/{tenant_slug}/go-live/rollouts/{public_id}/rollback")
def rollback_go_live_rollout(tenant_slug: str, public_id: str, payload: dict | None = None) -> dict:
    return _transition_go_live(tenant_slug, public_id, "rollback", payload)


@app.get("/api/platform-control/tenants/{tenant_slug}/lifecycle/jobs")
def lifecycle_jobs(tenant_slug: str, cursor: str | None = None, limit: int = 50) -> dict:
    return list_lifecycle_jobs_page(tenant_slug, cursor, limit)


@app.get("/api/platform-control/tenants/{tenant_slug}/lifecycle/jobs/{public_id}")
def lifecycle_job_detail(tenant_slug: str, public_id: str) -> dict:
    job = get_lifecycle_job(tenant_slug, public_id)
    if job is None:
        raise HTTPException(status_code=404, detail={"code": "lifecycle_job_not_found", "message": "Lifecycle job was not found."})
    return job


def _queue_lifecycle_job(tenant_slug: str, job_type: str, payload: dict, idempotency_key: str | None) -> JSONResponse:
    try:
        job = create_lifecycle_job(tenant_slug, job_type, payload, idempotency_key)
    except ValueError as error:
        raise HTTPException(status_code=400, detail={"code": str(error), "message": f"{job_type.capitalize()} payload is invalid."}) from error

    location = f"/api/platform-control/tenants/{tenant_slug}/lifecycle/jobs/{job['publicId']}"
    return JSONResponse(status_code=202, content=job, headers={"Location": location})


@app.post("/api/platform-control/tenants/{tenant_slug}/lifecycle/onboarding")
def onboarding(tenant_slug: str, payload: dict, idempotency_key: str | None = Header(default=None, alias="Idempotency-Key")) -> JSONResponse:
    return _queue_lifecycle_job(tenant_slug, "onboarding", payload, idempotency_key)


@app.post("/api/platform-control/tenants/{tenant_slug}/lifecycle/onboarding/preview")
def onboarding_preview(tenant_slug: str, payload: dict) -> dict:
    return build_lifecycle_preview(tenant_slug, "onboarding", payload)


@app.post("/api/platform-control/tenants/{tenant_slug}/lifecycle/offboarding")
def offboarding(tenant_slug: str, payload: dict, idempotency_key: str | None = Header(default=None, alias="Idempotency-Key")) -> JSONResponse:
    return _queue_lifecycle_job(tenant_slug, "offboarding", payload, idempotency_key)


@app.post("/api/platform-control/tenants/{tenant_slug}/lifecycle/offboarding/preview")
def offboarding_preview(tenant_slug: str, payload: dict) -> dict:
    return build_lifecycle_preview(tenant_slug, "offboarding", payload)


def _transition_lifecycle_job(tenant_slug: str, public_id: str, action: str, payload: dict) -> dict:
    try:
        return transition_lifecycle_job(tenant_slug, public_id, action, payload)
    except ValueError as error:
        if str(error) == "lifecycle_job_not_found":
            raise HTTPException(status_code=404, detail={"code": str(error), "message": "Lifecycle job was not found."}) from error
        raise HTTPException(status_code=400, detail={"code": str(error), "message": "Lifecycle transition payload is invalid."}) from error


@app.post("/api/platform-control/tenants/{tenant_slug}/lifecycle/jobs/{public_id}/start")
def start_lifecycle_job(tenant_slug: str, public_id: str, payload: dict | None = None) -> dict:
    return _transition_lifecycle_job(tenant_slug, public_id, "start", payload or {})


@app.post("/api/platform-control/tenants/{tenant_slug}/lifecycle/jobs/{public_id}/complete")
def complete_lifecycle_job(tenant_slug: str, public_id: str, payload: dict | None = None) -> dict:
    return _transition_lifecycle_job(tenant_slug, public_id, "complete", payload or {})


@app.post("/api/platform-control/tenants/{tenant_slug}/lifecycle/jobs/{public_id}/fail")
def fail_lifecycle_job(tenant_slug: str, public_id: str, payload: dict | None = None) -> dict:
    return _transition_lifecycle_job(tenant_slug, public_id, "fail", payload or {})


@app.post("/api/platform-control/tenants/{tenant_slug}/lifecycle/jobs/{public_id}/cancel")
def cancel_lifecycle_job(tenant_slug: str, public_id: str, payload: dict | None = None) -> dict:
    return _transition_lifecycle_job(tenant_slug, public_id, "cancel", payload or {})